Move ThreadSafeDeque into its own header to avoid polluting all other files
